+++ contenedores/src/main/java/backend/grupo130/contenedores/dto/request/EditRequest.java
@@ -23,6 +23,6 @@ public class EditRequest {
     private final BigDecimal volumen;
 
     @Positive(message = "El ID del cliente debe ser un número positivo")
-    private final Integer idCliente;
+    private final Long idCliente;
 
 }
