+++ contenedores/src/main/java/backend/grupo130/contenedores/dto/response/EditResponse.java
@@ -13,6 +13,6 @@ public class EditResponse {
 
     private final BigDecimal volumen;
 
-    private final Integer idCliente;
+    private final Long idCliente;
 
 }
