+++ contenedores/src/main/java/backend/grupo130/contenedores/repository/UsuarioRepository.java
@@ -15,7 +15,7 @@ public class UsuarioRepository {
 
     private final UsuarioClient usuarioRepository;
 
-    public Usuario getById(Integer usuarioId){
+    public Usuario getById(Long usuarioId){
 
         try {
